Move abbreviate tests into their own module.
